Added a new "Visual" class that represents just the information we need to track for an Effect once it's instantiated. Using it in place of Effect saves a string copy and a good deal of memory for each active visual effect.

+++ source/Body.h
@@ -103,7 +103,7 @@ protected:
 	Angle angle;
 	// A zoom of 1 means the sprite should be drawn at half size. For objects
 	// whose sprites should be full size, use zoom = 2.
-	double zoom = 1.;
+	float zoom = 1.f;
 	
 	// Government, for use in collision checks.
 	const Government *government = nullptr;
